+++ packages/ic-asset-certification/src/asset_router.rs
@@ -563,6 +563,7 @@ impl<'content> AssetRouter<'content> {
                     encoding,
                     Some(range_begin),
                 )?;
+                self.tree.borrow_mut().insert(&response.tree_entry);
                 self.responses.insert(
                     RequestKey::new(&asset_url, encoding_str(encoding), Some(range_begin)),
                     response,
@@ -761,6 +762,7 @@ impl<'content> AssetRouter<'content> {
             status_code,
             Cow::Owned(vec![]),
             headers,
+            vec![],
         )?;
 
         Ok(CertifiedAssetResponse {
@@ -792,6 +794,7 @@ impl<'content> AssetRouter<'content> {
             headers.push(("content-encoding".to_string(), encoding.to_string()));
         }
 
+        let mut request_headers = vec![];
         if let Some(range_begin) = range_begin {
             let total_length = content.len();
             let range_end = cmp::min(range_begin + ASSET_CHUNK_SIZE, total_length) - 1;
@@ -801,6 +804,7 @@ impl<'content> AssetRouter<'content> {
                 http::header::CONTENT_RANGE.to_string(),
                 format!("bytes {range_begin}-{range_end}/{total_length}"),
             ));
+            request_headers.push(http::header::RANGE.to_string());
         };
 
         Self::prepare_response_and_certification(
@@ -808,6 +812,7 @@ impl<'content> AssetRouter<'content> {
             status_code,
             content,
             headers,
+            request_headers,
         )
     }
 
@@ -815,13 +820,19 @@ impl<'content> AssetRouter<'content> {
         url: String,
         status_code: u16,
         body: Cow<'content, [u8]>,
-        additional_headers: Vec<(String, String)>,
+        additional_response_headers: Vec<(String, String)>,
+        certified_request_headers: Vec<String>,
     ) -> AssetCertificationResult<(HttpResponse<'content>, HttpCertification)> {
         let mut headers = vec![("content-length".to_string(), body.len().to_string())];
 
-        headers.extend(additional_headers);
-
+        headers.extend(additional_response_headers);
         let cel_expr = DefaultCelBuilder::full_certification()
+            .with_request_headers(
+                certified_request_headers
+                    .iter()
+                    .map(|s| s.as_str())
+                    .collect::<Vec<&str>>(),
+            )
             .with_response_certification(DefaultResponseCertification::response_header_exclusions(
                 vec![],
             ))
@@ -1123,7 +1134,7 @@ mod tests {
         let request = HttpRequest::get(&req_url).build();
         let mut expected_response = build_206_response(
             asset_body[0..ASSET_CHUNK_SIZE].to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -1170,7 +1181,7 @@ mod tests {
             let expected_range_end = cmp::min(asset_len_so_far + ASSET_CHUNK_SIZE, asset_len) - 1;
             let mut expected_response = build_206_response(
                 asset_body[asset_len_so_far..=expected_range_end].to_vec(),
-                asset_cel_expr(),
+                asset_range_chunk_cel_expr(),
                 vec![
                     (
                         "cache-control".to_string(),
@@ -1323,7 +1334,7 @@ mod tests {
             .build();
         let mut expected_response = build_206_response(
             asset_body[0..ASSET_CHUNK_SIZE].to_vec(),
-            encoded_asset_cel_expr(),
+            encoded_range_chunk_asset_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -1376,7 +1387,7 @@ mod tests {
             let expected_range_end = cmp::min(asset_len_so_far + ASSET_CHUNK_SIZE, asset_len) - 1;
             let mut expected_response = build_206_response(
                 asset_body[asset_len_so_far..=expected_range_end].to_vec(),
-                asset_cel_expr(),
+                encoded_range_chunk_asset_cel_expr(),
                 vec![
                     (
                         "cache-control".to_string(),
@@ -2789,7 +2800,7 @@ mod tests {
                 .get(format!("/{}", TWO_CHUNKS_ASSET_NAME), None, Some(0));
         let expected_first_chunk_response = build_206_response(
             first_chunk_body.to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -2815,7 +2826,7 @@ mod tests {
         );
         let expected_first_chunk_gzip_response = build_206_response(
             first_chunk_gzip_body.to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -2846,7 +2857,7 @@ mod tests {
         );
         let expected_second_chunk_response = build_206_response(
             second_chunk_body.to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -2877,7 +2888,7 @@ mod tests {
         );
         let expected_second_chunk_gzip_response = build_206_response(
             second_chunk_gzip_body.to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -2952,10 +2963,12 @@ mod tests {
         let assets: Vec<_> = asset_router.get_assets().iter().collect();
         assert!(assets.len() == 3);
 
+        println!("{:#?}", assets);
+
         let first_chunk_body = &full_body[0..ASSET_CHUNK_SIZE];
         let expected_first_chunk_response = build_206_response(
             first_chunk_body.to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -2968,6 +2981,8 @@ mod tests {
                 ),
             ],
         );
+
+        println!("expected first chunk: {:#?}", expected_first_chunk_response);
         assert!(assets.contains(&(
             (&format!("/{}", TWO_CHUNKS_ASSET_NAME), None, Some(0)),
             &expected_first_chunk_response
@@ -2976,7 +2991,7 @@ mod tests {
         let second_chunk_body = &full_body[ASSET_CHUNK_SIZE..full_body.len()];
         let expected_second_chunk_response = build_206_response(
             second_chunk_body.to_vec(),
-            asset_cel_expr(),
+            asset_range_chunk_cel_expr(),
             vec![
                 (
                     "cache-control".to_string(),
@@ -3493,8 +3508,30 @@ mod tests {
     }
 
     #[fixture]
+    fn asset_range_chunk_cel_expr() -> String {
+        DefaultFullCelExpressionBuilder::default()
+            .with_request_headers(vec!["range"])
+            .with_response_certification(DefaultResponseCertification::response_header_exclusions(
+                vec![],
+            ))
+            .build()
+            .to_string()
+    }
+
+    #[fixture]
     fn encoded_asset_cel_expr() -> String {
         DefaultFullCelExpressionBuilder::default()
+            .with_response_certification(DefaultResponseCertification::response_header_exclusions(
+                vec![],
+            ))
+            .build()
+            .to_string()
+    }
+
+    #[fixture]
+    fn encoded_range_chunk_asset_cel_expr() -> String {
+        DefaultFullCelExpressionBuilder::default()
+            .with_request_headers(vec!["range"])
             .with_response_certification(DefaultResponseCertification::response_header_exclusions(
                 vec![],
             ))